2885: Small fixes
