+++ web/src/components/PoiChips.tsx
@@ -1,0 +1,66 @@
+import styled from '@emotion/styled'
+import React, { ReactElement } from 'react'
+import { useTranslation } from 'react-i18next'
+
+import { PoiModel } from 'shared/api'
+
+import { AccessibleIcon, NotAccessibleIcon } from '../assets'
+import Icon from './base/Icon'
+
+const ChipsContainer = styled.div`
+  display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
+  margin-top: 8px;
+  margin-bottom: 8px;
+  align-items: center;
+  justify-content: flex-start;
+`
+
+const Chip = styled.div`
+  color: ${props => props.theme.colors.textSecondaryColor};
+  display: flex;
+  padding-inline: 12px;
+  align-items: center;
+  gap: 6px;
+  border-radius: 12px;
+  border: 1px solid rgb(0 0 0 / 38%);
+  height: 20px;
+  font-size: 12px;
+`
+
+const ChipIcon = styled(Icon)`
+  width: 12px;
+  height: 12px;
+`
+
+const PoiChips = ({ poi }: { poi: PoiModel }): ReactElement => {
+  const { t } = useTranslation()
+  const { category } = poi
+
+  const barrierFreeChip =
+    poi.barrierFree === true ? (
+      <>
+        <ChipIcon src={AccessibleIcon} />
+        {t('common:accessible')}
+      </>
+    ) : (
+      <>
+        <ChipIcon src={NotAccessibleIcon} />
+        {t('common:notAccessible')}
+      </>
+    )
+
+  return (
+    <ChipsContainer>
+      {poi.barrierFree !== null && <Chip>{barrierFreeChip}</Chip>}
+      {poi.organization !== null && <Chip>{poi.organization.name}</Chip>}
+      <Chip>
+        <ChipIcon src={category.icon} />
+        {category.name}
+      </Chip>
+    </ChipsContainer>
+  )
+}
+
+export default PoiChips
