+++ web/src/components/PoiDetails.tsx
@@ -6,19 +6,14 @@ import { useTranslation } from 'react-i18next'
 import { getExternalMapsLink } from 'shared'
 import { PoiModel } from 'shared/api'
 
-import {
-  AccessibleIcon,
-  ExternalLinkIcon,
-  LocationIcon,
-  NotAccessibleIcon,
-  PoiThumbnailPlaceholderLarge,
-} from '../assets'
+import { ExternalLinkIcon, LocationIcon, PoiThumbnailPlaceholderLarge } from '../assets'
 import dimensions from '../constants/dimensions'
 import { helpers } from '../constants/theme'
 import useWindowDimensions from '../hooks/useWindowDimensions'
 import Collapsible from './Collapsible'
 import Contact from './Contact'
 import OpeningHours from './OpeningHours'
+import PoiChips from './PoiChips'
 import RemoteContent from './RemoteContent'
 import Spacer from './Spacer'
 import Icon from './base/Icon'
@@ -55,33 +50,6 @@ const Thumbnail = styled.img`
 
 const Distance = styled.div`
   ${helpers.adaptiveFontSize};
-`
-
-const ChipsContainer = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-  margin-top: 8px;
-  margin-bottom: 8px;
-  align-items: center;
-  justify-content: flex-start;
-`
-
-const Chip = styled.div`
-  color: ${props => props.theme.colors.textSecondaryColor};
-  display: flex;
-  padding-inline: 12px;
-  align-items: center;
-  gap: 6px;
-  border-radius: 12px;
-  border: 1px solid rgb(0 0 0 / 38%);
-  height: 20px;
-  font-size: 12px;
-`
-
-const ChipIcon = styled(Icon)`
-  width: 12px;
-  height: 12px;
 `
 
 const AddressContentWrapper = styled.div`
@@ -161,8 +129,7 @@ const PoiDetails = ({ poi, distance, toolbar }: PoiDetailsProps): ReactElement =
   const { viewportSmall } = useWindowDimensions()
   const theme = useTheme()
   const { t } = useTranslation('pois')
-  const { content, location, contacts, isCurrentlyOpen, openingHours, temporarilyClosed, category, appointmentUrl } =
-    poi
+  const { content, location, contacts, isCurrentlyOpen, openingHours, temporarilyClosed, appointmentUrl } = poi
 
   const thumbnail = poi.thumbnail ?? PoiThumbnailPlaceholderLarge
   const isAndroid = /Android/i.test(navigator.userAgent)
@@ -170,33 +137,13 @@ const PoiDetails = ({ poi, distance, toolbar }: PoiDetailsProps): ReactElement =
   const appointmentOverlayUrl = appointmentUrl ?? poi.contacts.find(contact => contact.website !== null)?.website ?? ''
   const isOnlyWithAppointment = !openingHours && !!appointmentOverlayUrl
 
-  const barrierFreeChip =
-    poi.barrierFree === true ? (
-      <>
-        <ChipIcon src={AccessibleIcon} />
-        {t('common:accessible')}
-      </>
-    ) : (
-      <>
-        <ChipIcon src={NotAccessibleIcon} />
-        {t('common:notAccessible')}
-      </>
-    )
-
   return (
     <DetailsContainer>
       <HeadingSection>
         <Thumbnail alt='' src={thumbnail} />
         <Heading>{poi.title}</Heading>
         {distance !== null && <Distance>{t('distanceKilometre', { distance: distance.toFixed(1) })}</Distance>}
-        <ChipsContainer>
-          {poi.barrierFree !== null && <Chip>{barrierFreeChip}</Chip>}
-          <Chip>{poi.organization?.name}</Chip>
-          <Chip>
-            <ChipIcon src={category.icon} />
-            {category.name}
-          </Chip>
-        </ChipsContainer>
+        <PoiChips poi={poi} />
       </HeadingSection>
       <Spacer borderColor={theme.colors.borderColor} />
       {!viewportSmall && <Subheading>{t('detailsAddress')}</Subheading>}
